requirements.txt
cryptography
fastapi[standard]
jinja2
imagination
psycopg2-binary
pyyaml~=6.0
pyjwt[crypto]
requests
sqlalchemy
sqlalchemy[asyncio]
